Added prefix to all resources to avoid name clashes
Aaaah, that was why.

After a silly amount of debugging and analyzing the generated
resources, the solution is pretty obvious, after all. This convenient
syntax we all use, `@+id` rewrites IDs.

And it so happens that we still have the code of the old toasts in
Pixable's code, so the IDs with the same name were just rewritten, and
ended up trying to access silly things.